+++ scripts/dataAccess.js
@@ -1,0 +1,20 @@
+import { database } from "./database.js"
+
+
+
+export const setFacility = (facilityId) => {
+    database.transientState.selectedFacility = facilityId
+    document.dispatchEvent( new CustomEvent("stateChanged") )
+}
+
+export const getFacilities = () => {
+    return database.facilities.map(f => ({...f}))
+}
+
+export const purchaseMineral = () => {
+
+        // Broadcast custom event to entire documement so that the
+        // application can re-render and update state
+        document.dispatchEvent( new CustomEvent("stateChanged") )
+}
+
