+++ scripts/database.js
@@ -1,4 +1,4 @@
-const database = {
+export const database = {
     governors: [
         { id: 1, name: "Joe Dirt", activeStatus: true, colonyId: 1 },
         { id: 2, name: "Buffalo Bill", activeStatus: true, colonyId: 2 },
@@ -35,21 +35,5 @@ const database = {
     ],
     colonyMinerals: [],
     transientState: {}
-}
-
-export const setFacility = (facilityId) => {
-    database.transientState.selectedFacility = facilityId
-    document.dispatchEvent( new CustomEvent("stateChanged") )
-}
-
-export const getFacilities = () => {
-    return database.facilities.map(f => ({...f}))
-}
-
-export const purchaseMineral = () => {
-
-        // Broadcast custom event to entire documement so that the
-        // application can re-render and update state
-        document.dispatchEvent( new CustomEvent("stateChanged") )
 }
